10.4 compatibility
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@998 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCMView.m
@@ -7207,10 +7207,10 @@ BOOL	lowRes = NO;
 	[self setNeedsDisplay:YES];
 }
 
--(void)setTag:(NSInteger)aTag{
+-(void)setTag:(long)aTag{
 	_tag = aTag;
 }
-- (NSInteger)tag{
+- (long)tag{
 	return _tag;
 }
 -(float)curWW{
